# Backend/openapi.yaml
...
openapi: 3.0.3
info:
  title: SocietyEase API
  description: |
    API specification for the SocietyEase backend (Flask). Authentication uses JWT Bearer tokens. Each API endpoint is mapped to its corresponding user story using the x-user-story extension for traceability between the implementation and the project requirements.

    ## User story mapping

    Every operation carries an `x-user-story` extension. The stories are:

    | ID | User story |
    |---|---|
    | **US-01** | As a **Treasurer**, I want to generate and track maintenance invoices for every flat, so that dues are billed consistently and I can see what is outstanding. |
    | **US-02** | As a **Resident**, I want to raise a complaint about a problem in my flat or the building, so that the society has a record of it and someone is accountable for fixing it. |
    | **US-03** | As a **Resident**, I want to track the status of my complaint from open to closed, so that I know whether it is being worked on without having to chase anyone. |
    | **US-04** | As a **Association manager**, I want to assign a complaint to a maintenance worker, so that the right person is responsible and the job reaches their queue. |
    | **US-05** | As a **Owner**, I want to see the invoices raised against the flat I own, so that I can keep track of what my tenant or I owe the society. |
...

    The same table is maintained in `docs/USER_STORIES.md`.
  version: 4.1.0
servers:
- url: http://localhost:5000
...
paths:
...
  /api/invoices/:
    get:
      tags:
      - Invoices
      summary: List invoices
      description: |-
        Lists invoices, most recent first. Any admin role (ADMIN, SYSTEM_ADMIN, TREASURER, COMMITTEE_MEMBER) sees every invoice; a resident sees only their own flat's. Before this or any other invoice endpoint reads the table, a scoped bulk update promotes any UNPAID invoice whose due date has passed to OVERDUE, so status filters are never stale. The query filters below are applied on top of role scoping — they can only narrow what a caller could already see. **User Story — US-05, US-06, US-18**

        Access: any authenticated, active user.
      x-user-story:
      - US-05
      - US-06
      - US-18
      security:
      - bearerAuth: []
      parameters:
      - name: q
        in: query
        schema:
          type: string
        description: Alias for `flat` — matches the flat number.
      - name: flat
        in: query
        schema:
          type: string
      - name: status
        in: query
        schema:
          type: string
          enum:
          - PAID
          - UNPAID
          - OVERDUE
      - name: month
        in: query
        schema:
          type: integer
      - name: year
        in: query
        schema:
          type: integer
      - name: apartment_id
        in: query
        schema:
          type: integer
      - name: min_amount
        in: query
        schema:
          type: number
      - name: max_amount
        in: query
        schema:
          type: number
      - name: from
        in: query
        schema:
          type: string
        description: Inclusive lower bound on due_date (YYYY-MM-DD).
      - name: to
        in: query
        schema:
          type: string
        description: Inclusive upper bound on due_date (YYYY-MM-DD).
      responses:
        '200':
          description: Invoices list, most recent first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Invoice'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Missing, invalid, or expired JWT bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: invoices_get_invoices
      x-required-role: any authenticated, active user
    post:
      tags:
      - Invoices
      summary: Create invoice
      description: |-
        Generates a single maintenance invoice for one apartment for a given month/year. **User Story — US-01**

        Access: ADMIN, SYSTEM_ADMIN or TREASURER.
      x-user-story: US-01
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceCreateRequest'
        description: JSON object. Unknown fields are ignored; blank optional values are stored as null.
      responses:
        '201':
          description: Invoice created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '400':
          description: apartment_id, month, year, or amount is missing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing, invalid, or expired JWT bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: invoices_create_invoice
      x-required-role: ADMIN, SYSTEM_ADMIN or TREASURER
  /api/invoices/bulk:
    post:
      tags:
      - Invoices
      summary: Bulk-generate invoices
      description: |-
        Generates an invoice of the given amount for every apartment in the society for the given month/year, in a single call. Apartments that already have an invoice for that month/year are skipped. **User Story — US-01**

        Access: ADMIN, SYSTEM_ADMIN or TREASURER.
      x-user-story: US-01
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceBulkCreateRequest'
        description: JSON object. Unknown fields are ignored; blank optional values are stored as null.
      responses:
        '201':
          description: Invoices generated for all eligible flats.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  flats:
                    type: array
                    items:
                      type: string
        '400':
          description: month, year, or amount is missing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing, invalid, or expired JWT bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: invoices_bulk_generate
      x-required-role: ADMIN, SYSTEM_ADMIN or TREASURER
  /api/invoices/pending:
    get:
      tags:
      - Invoices
      summary: List all pending invoices
      description: |-
        Lists unpaid invoices (UNPAID or OVERDUE). Admins and the treasurer see every flat; a resident sees only their own flat's outstanding dues. Accepts the same filters as GET /api/invoices/, minus `status` (this endpoint already excludes PAID). **User Story — US-01, US-18**

        Access: any authenticated, active user.
      x-user-story:
      - US-01
      - US-18
      security:
      - bearerAuth: []
      parameters:
      - name: q
        in: query
        schema:
          type: string
      - name: flat
        in: query
        schema:
          type: string
      - name: month
        in: query
        schema:
          type: integer
      - name: year
        in: query
        schema:
          type: integer
      - name: apartment_id
        in: query
        schema:
          type: integer
      - name: min_amount
        in: query
        schema:
          type: number
      - name: max_amount
        in: query
        schema:
          type: number
      - name: from
        in: query
        schema:
          type: string
      - name: to
        in: query
        schema:
          type: string
      responses:
        '200':
          description: Pending invoices, across all apartments.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Invoice'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Missing, invalid, or expired JWT bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: invoices_get_pending
      x-required-role: any authenticated, active user
  /api/invoices/{inv_id}/pay:
    put:
      tags:
      - Invoices
      summary: Mark invoice paid
      description: |-
        Marks an invoice as PAID and records a corresponding payment/receipt entry against the resident of its apartment. **User Story — US-01**

        Access: ADMIN, SYSTEM_ADMIN or TREASURER.
      x-user-story: US-01
      security:
      - bearerAuth: []
      parameters:
      - name: inv_id
        in: path
        required: true
        schema:
          type: integer
        description: Invoice id.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
        description: JSON object. Unknown fields are ignored; blank optional values are stored as null.
      responses:
        '200':
          description: Invoice marked as paid; receipt generated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  invoice:
                    $ref: '#/components/schemas/Invoice'
                  receipt:
                    type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Missing, invalid, or expired JWT bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: No invoice with this id, or no resident is linked to the invoice's apartment.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: invoices_mark_paid
      x-required-role: ADMIN, SYSTEM_ADMIN or TREASURER
  /api/invoices/{inv_id}/receipt:
    get:
      tags:
      - Invoices
      summary: Get payment receipt
      description: |-
        Returns the payment receipt for an invoice. The invoice must already be marked PAID. **User Story — US-06**

        Access: any authenticated, active user.
      x-user-story: US-06
      security:
      - bearerAuth: []
      parameters:
      - name: inv_id
        in: path
        required: true
        schema:
          type: integer
        description: Invoice id.
      responses:
        '200':
          description: Payment receipt.
          content:
            application/json:
              schema:
                type: object
        '400':
          description: The invoice has not been paid yet.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing, invalid, or expired JWT bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: No invoice with this id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      operationId: invoices_get_receipt
      x-required-role: any authenticated, active user
  /api/invoices/summary:
    get:
      tags:
      - Invoices
      summary: Invoice totals, counts and collection rate
      description: |-
        Totals, PAID/UNPAID/OVERDUE counts and a monthly breakdown over the same role-scoped, optionally filtered set of invoices as GET /api/invoices/. Accepts the identical query filters. The overdue sweep runs first, so `count_overdue` is always current. **User Story — US-19**

        Access: any authenticated, active user (scoped exactly like the list).
      x-user-story: US-19
      security:
      - bearerAuth: []
      parameters:
      - name: q
        in: query
        schema:
          type: string
      - name: flat
        in: query
        schema:
          type: string
      - name: status
        in: query
        schema:
          type: string
      - name: month
        in: query
        schema:
          type: integer
      - name: year
        in: query
        schema:
          type: integer
      - name: apartment_id
        in: query
        schema:
          type: integer
      - name: min_amount
        in: query
        schema:
          type: number
      - name: max_amount
        in: query
        schema:
          type: number
      - name: from
        in: query
        schema:
          type: string
      - name: to
        in: query
        schema:
          type: string
      responses:
        '200':
          description: Invoice summary for the caller's scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
      operationId: invoices_summary
      x-required-role: any authenticated, active user
  /api/invoices/export:
    get:
      tags:
      - Invoices
      summary: Export invoices as CSV
      description: |-
        Same role scoping and query filters as GET /api/invoices/, returned as a downloadable CSV. **User Story — US-19**

        Access: any authenticated, active user (scoped exactly like the list).
      x-user-story: US-19
      security:
      - bearerAuth: []
      parameters:
      - name: q
        in: query
        schema:
          type: string
      - name: flat
        in: query
        schema:
          type: string
      - name: status
        in: query
        schema:
          type: string
      - name: month
        in: query
        schema:
          type: integer
      - name: year
        in: query
        schema:
          type: integer
      - name: apartment_id
        in: query
        schema:
          type: integer
      - name: min_amount
        in: query
        schema:
          type: number
      - name: max_amount
        in: query
        schema:
          type: number
      - name: from
        in: query
        schema:
          type: string
      - name: to
        in: query
        schema:
          type: string
      responses:
        '200':
          description: CSV file of the filtered invoices list.
          content:
            text/csv:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
      operationId: invoices_export
      x-required-role: any authenticated, active user
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
      - error
...
    Invoice:
      type: object
      properties:
        id:
          type: integer
        apartment_id:
          type: integer
        flat_number:
          type: string
        month:
          type: integer
        year:
          type: integer
        amount:
          type: number
          format: float
        due_date:
          type: string
          nullable: true
        status:
          type: string
          enum:
          - PAID
          - UNPAID
          - OVERDUE
        created_at:
          type: string
    InvoiceCreateRequest:
      type: object
      properties:
        apartment_id:
          type: integer
        month:
          type: integer
        year:
          type: integer
        amount:
          type: number
        due_date:
          type: string
      required:
      - apartment_id
      - month
      - year
      - amount
    InvoiceBulkCreateRequest:
      type: object
      properties:
        month:
          type: integer
        year:
          type: integer
        amount:
          type: number
        due_date:
          type: string
      required:
      - month
      - year
      - amount
    PaymentRequest:
      type: object
      properties:
        payment_method:
          type: string
        transaction_reference:
          type: string
...
    InvoiceSummary:
      type: object
      description: Totals and a monthly breakdown over the same role-scoped, optionally filtered set of invoices as GET /api/invoices/. The overdue sweep runs first, so counts are always current.
      properties:
        total_invoiced:
          type: number
        total_collected:
          type: number
        total_pending:
          type: number
        count_paid:
          type: integer
        count_unpaid:
          type: integer
        count_overdue:
          type: integer
        overdue_amount:
          type: number
        collection_rate:
          type: number
          description: total_collected / total_invoiced * 100, rounded to 2 decimals. 0 if nothing is invoiced.
        by_month:
          type: object
          description: 'Keyed by "YYYY-MM", each value an object with `invoiced` and `collected` totals.'
...
  responses:
    BadRequest:
      description: Validation failed — a required field is missing, or a value has the wrong type/format
        (bad date, non-numeric amount, unknown enum value, or a body that is not a JSON object).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: 'category must be one of: PLUMBING, ELECTRICAL, CLEANING, SECURITY, OTHER'
    Unauthorized:
      description: 'No JWT was supplied, or it is invalid or malformed. Note: flask-jwt-extended emits
        these under a `msg` key rather than `error`.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            msg: Missing Authorization Header
    Forbidden:
      description: 'The caller is authenticated but not allowed: either the account has been deactivated,
        or the user''s role is not permitted to perform this action, or the resource belongs to another
        resident.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: You are not allowed to perform this action
    NotFound:
      description: The requested resource does not exist. Also returned when the JWT is valid but the
        user it identifies has been removed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Not found
    Conflict:
      description: The request conflicts with existing data — a duplicate unique value, or a business
        rule such as paying an already-paid invoice.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: This invoice is already paid
    ServerError:
      description: Unexpected server error. All errors are returned as JSON; the API never returns an
        HTML error page.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Internal server error
...